Rollback some unneeded-stuff.
