+++ java/arcs/core/storage/driver/BUILD
@@ -7,21 +7,10 @@ licenses(["notice"])
 
 package(default_visibility = ["//visibility:public"])
 
-STORAGE_KEY_IMPL_SRC = [
-    "DatabaseStorageKey.kt",
-    "RamDiskStorageKey.kt",
-    "VolatileStorageKey.kt",
-]
-
 arcs_kt_library(
     name = "driver",
-    srcs = glob(
-        ["*.kt"],
-        exclude = STORAGE_KEY_IMPL_SRC,
-    ),
-    exports = [":storage_key_impl"],
+    srcs = glob(["*.kt"]),
     deps = [
-        ":storage_key_impl",
         "//java/arcs/core/common",
         "//java/arcs/core/crdt",
         "//java/arcs/core/crdt/extension",
@@ -35,14 +24,5 @@ arcs_kt_library(
         "//java/arcs/core/util",
         "//third_party/kotlin/kotlinx_atomicfu",
         "//third_party/kotlin/kotlinx_coroutines",
-    ],
-)
-
-arcs_kt_library(
-    name = "storage_key_impl",
-    srcs = STORAGE_KEY_IMPL_SRC,
-    deps = [
-        "//java/arcs/core/common",
-        "//java/arcs/core/storage:storage_key",
     ],
 )
